Refactor component file structure
Reorganized the project's file structure for better readability and maintainability. Components have been moved from general folders like 'atoms', 'molecules', and 'organisms' to more specific ones named after the components themselves, such as 'ui', 'sidebar', and 'footer'. This change affects multiple files across the application, including but not limited to ButtonPrimary, SignOutButton, Sidebar, and Footer.
